realignment some loc roomcontroller and apartmentcontroller

+++ app/Http/Controllers/RoomController.php
@@ -3,22 +3,22 @@
 namespace App\Http\Controllers;
 
 use App\Guestarrive;
-use App\Houseprice;
 use App\Models\District;
 use App\Models\Food;
 use App\Models\Himage;
 use App\Models\House;
 use App\Models\Houseamenity;
 use App\Models\Housedetail;
+use App\Models\Houseprice;
 use App\Models\Houserule;
 use App\Models\Housespace;
 use App\Models\Housetype;
 use App\Models\Map;
 use App\Models\Province;
+use App\Models\Review;
 use App\Models\SubDistrict;
 use App\Payment;
 use App\Rental;
-use App\Review;
 use App\User;
 use File;
 use Illuminate\Http\Request;
@@ -30,13 +30,42 @@ use Storage;
 
 class RoomController extends Controller
 {
+    private $select_types_global = ['type 2 apartment', 'type 3 apartment'];
+    /*publish flag 0 private, 1 public, 2 trash, 3 permanant delete*/
+
     public function __construct()
     {
         $this->middleware('auth', ['except' => ['index', 'show']]);
-        $this->middleware('crole:Admin')->except('index', 'create', 'show', 'store', 'edit', 'update', 'index_myroom');
+        $this->middleware('crole:Admin')->except('index', 'create', 'show', 'store', 'edit', 'update', 'destroy', 'index_myroom', 'owner');
     }
 
-    /*publish flag 0 private, 1 public, 2 trash, 3 permanant delete*/
+    private function getTypeId($request)
+    {
+        $select_types = $this->select_types_global;
+        if ($request == 'room') {
+            $types = Housetype::whereNotIn('name', $select_types)->get();
+        }
+        else {
+            $types = Housetype::whereIn('name', $select_types)->get();
+        }
+        $types_id = array();
+        foreach ($types as $key => $type) {
+            array_push($types_id, $type->id);
+        }
+        return $types_id;
+    }
+
+    private function getType($request)
+    {
+        $select_types = $this->select_types_global;
+        if ($request == 'room') {
+            $types = Housetype::whereNotIn('name', $select_types)->get();
+        }
+        else {
+            $types = Housetype::whereIn('name', $select_types)->get();
+        }
+        return $types;
+    }
 
     /**
      * Display a listing of the resource.
@@ -51,7 +80,8 @@ class RoomController extends Controller
 
     public function index_myroom(User $user) {
         if (Auth::user()->id === $user->id) {
-            $houses = House::where('users_id', $user->id)->where('housetypes_id','!=','2')->where('housetypes_id','!=','3')->orderBy('id')->paginate(10);
+            $types_id = $this->getTypeId('room');
+            $houses = House::where('users_id', $user->id)->whereIn('housetypes_id', $types_id)->orderBy('id')->paginate(10);
             return view('rooms.index-myroom')->with('houses', $houses);
         }
         Session::flash('fail', 'Unauthorized access.');
@@ -65,7 +95,7 @@ class RoomController extends Controller
      */
     public function create()
     {
-        $types = Housetype::where('name', '!=', 'Apartment')->where('name', '!=', 'Hotel')->get();
+        $types = $this->getType('room');
         $provinces = Province::all();
         $districts = District::where('province_id', $provinces[0]->id)->get();
         $sub_districts = SubDistrict::where('district_id', $districts[0]->id)->get();
@@ -180,9 +210,12 @@ class RoomController extends Controller
             foreach ($request->image_names as $image_name) {
                 $images = new Himage;
                 $filename = time() . rand(9,99) . Auth::user()->id . '.' . $image_name->getClientOriginalExtension();
-                $location = public_path('images/houses/'.$filename);
+                $location = public_path('images/houses/'.$house->id.'/');
+                if (!file_exists($location)) {
+                    $result = File::makeDirectory($location, 0775, true);
+                }
+                $location = public_path('images/houses/'.$house->id.'/'.$filename);
                 Image::make($image_name)->resize(1440, 1080)->save($location);
-
                 $images->houses_id = $house->id;
                 $images->image_name = $filename;
                 $images->save();
@@ -201,11 +234,13 @@ class RoomController extends Controller
         $house->housedetails()->sync($request->housedetails, false);
 
         Session::flash('success', 'This house was succussfully saved!');
-        return redirect()->route('rooms.single', $house->id);
+        return redirect()->route('rooms.owner', $house->id);
     }
 
-    public function owner(House $house){
-        if ($house->housetypes_id !== '2' || $house->housetypes_id !== '3') {
+    public function owner($houseId){
+        $types_id = $this->getTypeId('room');
+        $house = House::where('id', $houseId)->whereIn('housetypes_id', $types_id)->first();
+        if (!is_null($house)) {
             if (Auth::user()->hasRole('Admin') || Auth::user()->id == $house->users_id) {
                 $rentcount = Rental::where('houses_id', $house->id)->count();
                 $images = Himage::where('houses_id', $house->id)->get();
@@ -215,9 +250,7 @@ class RoomController extends Controller
             Session::flash('fail', 'Unauthorized access.');
             return back();
         }
-        else {
-            return redirect()->route('apartments.owner', $house->id);
-        }
+        return redirect()->route('apartments.owner', $houseId);
     }
 
     /**
@@ -226,34 +259,27 @@ class RoomController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function show(House $room)
+    public function show($houseId)
     {
-        $house = $room;
-        if (is_null($house)) {
-            Session::flash('fail', 'This room is no longer available.');
-            return back();
-        }
-        else {
+        $types_id = $this->getTypeId('room');
+        $house = House::where('id', $houseId)->whereIn('housetypes_id', $types_id)->first();
+        if (!is_null($house)) {
             if ($house->publish == '1') {
-                if ($house->housetypes_id == '2' || $house->housetypes_id == '3') {
-                    return redirect()->route('apartments.show', $id);
-                }
-                else {
-                    $images = Himage::where('houses_id', $house->id)->get();
-                    $reviews = Review::where('house_id', $house->id)->get();
-                    $clean = Review::where('house_id', $house->id)->avg('clean');
-                    $amenity = Review::where('house_id', $house->id)->avg('amenity');
-                    $service = Review::where('house_id', $house->id)->avg('service');
-                    $host = Review::where('house_id', $house->id)->avg('host');
-                    $avg = ($clean + $amenity + $service + $host)/4;
-                    $avg = number_format((float)$avg, 2, '.', '');
-                    $map = Map::where('houses_id', $house->id)->first();
-                    return view('rooms.show')->with('house', $house)->with('images', $images)->with('avg', $avg)->with('map', $map);
-                }
+                $images = Himage::where('houses_id', $house->id)->get();
+                $reviews = Review::where('house_id', $house->id)->get();
+                $clean = Review::where('house_id', $house->id)->avg('clean');
+                $amenity = Review::where('house_id', $house->id)->avg('amenity');
+                $service = Review::where('house_id', $house->id)->avg('service');
+                $host = Review::where('house_id', $house->id)->avg('host');
+                $avg = ($clean + $amenity + $service + $host)/4;
+                $avg = number_format((float)$avg, 2, '.', '');
+                $map = Map::where('houses_id', $house->id)->first();
+                return view('rooms.show')->with('house', $house)->with('images', $images)->with('avg', $avg)->with('map', $map);
             }
             Session::flash('fail', 'This room is no longer available.');
             return back();
         }
+        return redirect()->route('apartments.show', $houseId);
     }
 
     /**
@@ -265,17 +291,15 @@ class RoomController extends Controller
     public function edit(House $room)
     {
         $house = $room;
-        if ($house->housetypes_id == '2' || $house->housetypes_id == '3') {
+        if ($house->housetypes_id == '2') {
+            if ($house->housetypes_id == '3') {
+                return redirect()->route('apartments.edit', $house->id);
+            }
             return redirect()->route('apartments.edit', $house->id);
         }
         else {
             if (Auth::user()->id == $house->users_id) {
-                $housetypes = Housetype::where('name', '!=', 'Apartment')->where('name', '!=', 'Hotel')->get();
-                $types = array();
-                foreach ($housetypes as $housetype) {
-                    $types[$housetype->id] = $housetype->name;
-                }
-
+                $types = $this->getType('room');
                 $provinces = Province::all();
                 $districts = District::where('province_id', $provinces[0]->id)->get();
                 $sub_districts = SubDistrict::where('district_id', $districts[0]->id)->get();
@@ -285,32 +309,11 @@ class RoomController extends Controller
                 if ($house->district_id !== null) {
                     $sub_districts = SubDistrict::where('district_id', $house->district_id)->get();
                 }
-
-                $houseamenities = Houseamenity::all();
-                $amenities = array();
-                foreach ($houseamenities as $houseamenity) {
-                    $amenities[$houseamenity->id] = $houseamenity->amenityname;
-                }
-
-                $housespaces = Housespace::all();
-                $spaces = array();
-                foreach ($housespaces as $housespace) {
-                    $spaces[$housespace->id] = $housespace->spacename;
-                }
-
+                $amenities = Houseamenity::all();
+                $spaces = Housespace::all();
                 $houseimages = Himage::where('houses_id', $house->id)->get();
-
-                $houserules = Houserule::all();
-                $rules = array();
-                foreach ($houserules as $houserule) {
-                    $rules[$houserule->id] = $houserule->houserule_name;
-                }
-
-                $housedetails = Housedetail::all();
-                $details = array();
-                foreach ($housedetails as $housedetail) {
-                    $details[$housedetail->id] = $housedetail->must_know;
-                }
+                $rules = Houserule::all();
+                $details = Housedetail::all();
                 return view('rooms.edit')->with('house', $house)->with('types', $types)->with('sub_districts', $sub_districts)->with('districts', $districts)->with('provinces', $provinces)->with('amenities', $amenities)->with('spaces', $spaces)->with('houseimages', $houseimages)->with('rules', $rules)->with('details', $details);
             }
             Session::flash('fail', 'Unauthorized access.');
@@ -400,7 +403,11 @@ class RoomController extends Controller
             foreach ($request->image_names as $image_name) {
                 $image = new Himage;
                 $filename = time() . rand(9,99) . Auth::user()->id . '.' . $image_name->getClientOriginalExtension();
-                $location = public_path('images/houses/'.$filename);
+                $location = public_path('images/houses/'.$house->id.'/');
+                if (!file_exists($location)) {
+                    $result = File::makeDirectory($location, 0775, true);
+                }
+                $location = public_path('images/houses/'.$house->id.'/'.$filename);
                 Image::make($image_name)->resize(1440, 1080)->save($location);
                 $image->houses_id = $house->id;
                 $image->image_name = $filename;
@@ -450,50 +457,53 @@ class RoomController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function destroy($id)
+    public function destroy($houseId)
     {
-        $house = House::find($id);
-        $rental = Rental::where('houses_id', $house->id)->first();
-        $images = Himage::all();
-        if ($rental == NULL){
-            $house->houseamenities()->detach();
-            $house->housespaces()->detach();
-            $house->houserules()->detach();
-            $house->housedetails()->detach();
-            foreach ($images as $image) {
-                if ($image->houses_id == $house->id) {
-                    $filename = $image->image_name;
-                    $image->image_name = $filename;
-                    $image->delete();
-                    $location = public_path('images/houses/'.$filename);
-                    File::delete($location);
+        $house = House::find($houseId);
+        if (Auth::user()->id == $house->users_id) {
+            $rental = Rental::where('houses_id', $house->id)->first();
+            $images = Himage::all();
+            if ($rental == NULL){
+                $house->houseamenities()->detach();
+                $house->housespaces()->detach();
+                $house->houserules()->detach();
+                $house->housedetails()->detach();
+                foreach ($images as $image) {
+                    if ($image->houses_id == $house->id) {
+                        $filename = $image->image_name;
+                        $image->image_name = $filename;
+                        $image->delete();
+                        $location = public_path('images/houses/'.$houseId.'/'.$filename);
+                        File::delete($location);
+                    }
                 }
+                $map = Map::where('houses_id', $house->id)->first();
+                $map->delete();
+                $house->delete();
+                $houseprice = Houseprice::find($house->houseprices_id);
+                $houseprice->delete();
+                $food = Food::find($house->foods_id);
+                $food->delete();
+                $guestarrive = Guestarrive::find($house->guestarrives_id);
+                $guestarrive->delete();
+                $alt = 'Room #ID ' . $house->id . ' Deleted';
             }
-            $map = Map::where('houses_id', $house->id)->first();
-            $map->delete();
-            $house->delete();
-            $houseprice = Houseprice::find($house->houseprices_id);
-            $houseprice->delete();
-            $food = Food::find($house->foods_id);
-            $food->delete();
-            $guestarrive = Guestarrive::find($house->guestarrives_id);
-            $guestarrive->delete();
-            $alt = 'Room #ID ' . $house->id . ' Deleted';
+            else {
+                $alt = 'Can not delete Room #ID ' . $house->id . ' because someone has rented.';
+            }
+            return redirect()->route('rooms.index-myroom', Auth::user()->id)->with('alert', $alt);
         }
-        else {
-            $alt = 'Can not delete Room #ID ' . $house->id . ' because someone has rented.';
-        }
-
-        return redirect()->route('index-myroom', Auth::user()->id)->with('alert', $alt);
+        Session::flash('fail', 'Unauthorized access.');
+        return back();
     }
 
     public function detroyimage(Himage $image)
     {
         $filename = $image->image_name;
-        $house_id = $image->houses_id;
-        $location = public_path('images/houses/'.$filename);
+        $houseId = $image->houses_id;
+        $location = public_path('images/houses/'.$houseId.'/'.$filename);
         File::delete($location);
         $image->delete();
-        return redirect()->route('rooms.owner', $house_id);
+        return redirect()->route('rooms.owner', $houseId);
     }
 }
